GRAPHICS: Fix potential one-definition-rule overlaps in mismatched-architecture modules.
