+++ common/scummsys.h
@@ -22,6 +22,13 @@
 #ifndef COMMON_SCUMMSYS_H
 #define COMMON_SCUMMSYS_H
 
+// WARNING: DO NOT PUT ANY FUNCTION DEFINITIONS (I.E. CODE) IN THIS FILE
+// ON ARCHITECTURES THAT COMPILE FILES WITH DIFFERENT INSTRUCTION SET
+// EXTENSIONS, LIKE ARM AND X86/X64 IN GCC/CLANG BUILDS.
+//
+// DO NOT INCLUDE ANYTHING THAT CONTAINS FUNCTION DEFINITIONS ON THOSE
+// PLATFORMS EITHER.
+
 #ifndef __has_feature           // Optional of course.
 	#define __has_feature(x) 0  // Compatibility with non-clang compilers.
 #endif
@@ -541,50 +548,13 @@ namespace std {
 }
 #endif
 
-//
-// std::initializer_list
-// Provide replacement when not available
-//
-#if defined(NO_CXX11_INITIALIZER_LIST)
-namespace std {
-	template<class T> class initializer_list {
-	public:
-		typedef T value_type;
-		typedef const T &reference;
-		typedef const T &const_reference;
-		typedef size_t size_type;
-		typedef const T *iterator;
-		typedef const T *const_iterator;
-
-		constexpr initializer_list() noexcept = default;
-		constexpr size_t size() const noexcept { return _size; };
-		constexpr const T *begin() const noexcept { return _begin; };
-		constexpr const T *end() const noexcept { return _begin + _size; }
-
-	private:
-		// Note: begin has to be first or the compiler may get very upset
-		const T *_begin = { nullptr };
-		size_t _size = { 0 };
-
-		// The compiler is allowed to call this constructor
-		constexpr initializer_list(const T* t, size_t s) noexcept : _begin(t) , _size(s) {}
-	};
-
-	template<class T> constexpr const T* begin(initializer_list<T> il) noexcept {
-		return il.begin();
-	}
-
-	template<class T> constexpr const T* end(initializer_list<T> il) noexcept {
-		return il.end();
-	}
-}
-
-#else
-
-#include <initializer_list>
-
-#endif // NO_CXX11_INITIALIZER_LIST
-
 #include "common/forbidden.h"
 
 #endif
+
+// WARNING: DO NOT PUT ANY FUNCTION DEFINITIONS (I.E. CODE) IN THIS FILE
+// ON ARCHITECTURES THAT COMPILE FILES WITH DIFFERENT INSTRUCTION SET
+// EXTENSIONS, LIKE ARM AND X86/X64 IN GCC/CLANG BUILDS.
+//
+// DO NOT INCLUDE ANYTHING THAT CONTAINS FUNCTION DEFINITIONS ON THOSE
+// PLATFORMS EITHER.
