+++ graphics/blit/blit-sse2.cpp
@@ -23,7 +23,6 @@
 #include <immintrin.h>
 
 #include "graphics/blit/blit-alpha.h"
-#include "graphics/pixelformat.h"
 
 namespace Graphics {
 
@@ -33,11 +32,11 @@ static FORCEINLINE __m128i sse2_mul32(__m128i a, __m128i b) {
 	return _mm_unpacklo_epi32(even, odd);
 }
 
-class BlendBlitImpl_SSE2 : public BlendBlitImpl_Base {
+class BlendBlitImpl_SSE2 : public BlendBlitImpl_Base<BlendBlitImpl_SSE2> {
 	friend class BlendBlit;
 
 template<bool doscale, bool rgbmod, bool alphamod>
-struct AlphaBlend : public BlendBlitImpl_Base::AlphaBlend<doscale, rgbmod, alphamod> {
+struct AlphaBlend : public BlendBlitImpl_Base<BlendBlitImpl_SSE2>::AlphaBlend<doscale, rgbmod, alphamod> {
 	static inline __m128i simd(__m128i src, __m128i dst, const bool flip, const byte ca, const byte cr, const byte cg, const byte cb) {
 		__m128i ina;
 		if (alphamod)
@@ -84,7 +83,7 @@ struct AlphaBlend : public BlendBlitImpl_Base::AlphaBlend<doscale, rgbmod, alpha
 };
 
 template<bool doscale, bool rgbmod, bool alphamod>
-struct MultiplyBlend : public BlendBlitImpl_Base::MultiplyBlend<doscale, rgbmod, alphamod> {
+struct MultiplyBlend : public BlendBlitImpl_Base<BlendBlitImpl_SSE2>::MultiplyBlend<doscale, rgbmod, alphamod> {
 	static inline __m128i simd(__m128i src, __m128i dst, const bool flip, const byte ca, const byte cr, const byte cg, const byte cb) {
 		__m128i ina, alphaMask;
 		if (alphamod) {
@@ -130,14 +129,14 @@ struct MultiplyBlend : public BlendBlitImpl_Base::MultiplyBlend<doscale, rgbmod,
 };
 
 template<bool doscale, bool rgbmod, bool alphamod>
-struct OpaqueBlend : public BlendBlitImpl_Base::OpaqueBlend<doscale, rgbmod, alphamod> {
+struct OpaqueBlend : public BlendBlitImpl_Base<BlendBlitImpl_SSE2>::OpaqueBlend<doscale, rgbmod, alphamod> {
 	static inline __m128i simd(__m128i src, __m128i dst, const bool flip, const byte ca, const byte cr, const byte cg, const byte cb) {
 		return _mm_or_si128(src, _mm_set1_epi32(BlendBlit::kAModMask));
 	}
 };
 
 template<bool doscale, bool rgbmod, bool alphamod>
-struct BinaryBlend : public BlendBlitImpl_Base::BinaryBlend<doscale, rgbmod, alphamod> {
+struct BinaryBlend : public BlendBlitImpl_Base<BlendBlitImpl_SSE2>::BinaryBlend<doscale, rgbmod, alphamod> {
 	static inline __m128i simd(__m128i src, __m128i dst, const bool flip, const byte ca, const byte cr, const byte cg, const byte cb) {
 		__m128i alphaMask = _mm_cmpeq_epi32(_mm_and_si128(src, _mm_set1_epi32(BlendBlit::kAModMask)), _mm_setzero_si128());
 		dst = _mm_and_si128(dst, alphaMask);
@@ -147,7 +146,7 @@ struct BinaryBlend : public BlendBlitImpl_Base::BinaryBlend<doscale, rgbmod, alp
 };
 
 template<bool doscale, bool rgbmod, bool alphamod>
-struct AdditiveBlend : public BlendBlitImpl_Base::AdditiveBlend<doscale, rgbmod, alphamod> {
+struct AdditiveBlend : public BlendBlitImpl_Base<BlendBlitImpl_SSE2>::AdditiveBlend<doscale, rgbmod, alphamod> {
 	static inline __m128i simd(__m128i src, __m128i dst, const bool flip, const byte ca, const byte cr, const byte cg, const byte cb) {
 		__m128i ina;
 		if (alphamod)
@@ -201,7 +200,7 @@ struct AdditiveBlend : public BlendBlitImpl_Base::AdditiveBlend<doscale, rgbmod,
 };
 
 template<bool doscale, bool rgbmod, bool alphamod>
-struct SubtractiveBlend : public BlendBlitImpl_Base::SubtractiveBlend<doscale, rgbmod, alphamod> {
+struct SubtractiveBlend : public BlendBlitImpl_Base<BlendBlitImpl_SSE2>::SubtractiveBlend<doscale, rgbmod, alphamod> {
 	static inline __m128i simd(__m128i src, __m128i dst, const bool flip, const byte ca, const byte cr, const byte cg, const byte cb) {
 		__m128i ina = _mm_and_si128(src, _mm_set1_epi32(BlendBlit::kAModMask));
 		__m128i srcb = _mm_srli_epi32(_mm_and_si128(src, _mm_set1_epi32(BlendBlit::kBModMask)), BlendBlit::kBModShift);
